refactor(core-lib): absorb precompile wrappers
