+++ Cargo.toml
@@ -24,7 +24,7 @@ members = [
     "crates/utils-sync",
     "miden-vm",
     "precompiles",
-    "precompiles/codegen",
+    "crates/lib/core/codegen",
     "processor",
     "prover",
     "verifier",
@@ -75,7 +75,7 @@ miden-debug-types       = { path = "./crates/debug-types", version = "0.25", def
 miden-mast-package      = { path = "./crates/mast-package", version = "0.25", default-features = false }
 miden-package-registry  = { path = "./crates/package-registry", version = "0.25", default-features = false }
 miden-precompiles       = { path = "./precompiles", version = "0.25", default-features = false }
-miden-precompiles-codegen = { path = "./precompiles/codegen", version = "0.25", default-features = false }
+miden-core-lib-codegen  = { path = "./crates/lib/core/codegen", version = "0.25", default-features = false }
 miden-processor         = { path = "./processor", version = "0.25", default-features = false }
 miden-test-serde-macros = { path = "./crates/test-serde-macros", default-features = false }
 miden-project           = { path = "./crates/project", version = "0.25", default-features = false }
