+++ crates/lib/core/codegen/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "miden-precompiles-codegen"
+name = "miden-core-lib-codegen"
 version.workspace = true
-description = "Build-time MASM code generation support for miden-precompiles"
+description = "Build-time MASM code generation support for miden-core-lib"
 license.workspace = true
 authors.workspace = true
 homepage.workspace = true
@@ -11,16 +11,17 @@ edition.workspace = true
 
 [features]
 default = ["std"]
-std = ["miden-core/std"]
+std = ["miden-core/std", "miden-precompiles/std"]
 
 [dependencies]
 miden-core.workspace = true
+miden-precompiles.workspace = true
 
 [lib]
 bench = false
 doctest = false
 
 [[bin]]
-name = "miden-precompiles-codegen"
+name = "miden-core-lib-codegen"
 path = "src/main.rs"
 required-features = ["std"]
