+++ crates/lib/core/codegen/src/lib.rs
@@ -5,9 +5,5 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
-pub mod descriptors;
-
-pub use descriptors::*;
-
 #[cfg(feature = "std")]
 pub mod masm;
