+++ crates/lib/core/codegen/src/main.rs
@@ -1,14 +1,16 @@
 use std::{path::PathBuf, process::ExitCode};
 
-use miden_precompiles_codegen::masm;
+use miden_core_lib_codegen::masm;
+
+const DEFAULT_OUT_DIR: &str = "target/miden-core-lib-generated-asm";
 
 fn usage() {
     println!("Usage:");
-    println!("  miden-precompiles-codegen [--out <dir>]");
+    println!("  miden-core-lib-codegen [--out <dir>]");
     println!();
     println!("Options:");
     println!("  --out <dir>  Write generated MASM preview to this directory");
-    println!("               [default: target/miden-precompiles-generated-asm]");
+    println!("               [default: {DEFAULT_OUT_DIR}]");
     println!("  -h, --help   Show this message");
 }
 
@@ -43,9 +45,7 @@ fn parse_out_dir() -> Result<Option<PathBuf>, ExitCode> {
 
 fn main() -> ExitCode {
     let out_dir = match parse_out_dir() {
-        Ok(out_dir) => {
-            out_dir.unwrap_or_else(|| PathBuf::from("target/miden-precompiles-generated-asm"))
-        },
+        Ok(out_dir) => out_dir.unwrap_or_else(|| PathBuf::from(DEFAULT_OUT_DIR)),
         Err(code) => return code,
     };
 
@@ -54,6 +54,6 @@ fn main() -> ExitCode {
         return ExitCode::FAILURE;
     }
 
-    println!("wrote generated precompile MASM to {}", out_dir.display());
+    println!("wrote generated core-lib MASM to {}", out_dir.display());
     ExitCode::SUCCESS
 }
