+++ crates/lib/core/codegen/src/templates/curve.masm.tpl
@@ -1,4 +1,4 @@
-# @generated by miden-precompiles. Do not edit by hand.
+# @generated by miden-core-lib-codegen. Do not edit by hand.
 # Source template: {{TEMPLATE_PATH}}
 # Regenerate with: {{REGENERATE_COMMAND}}
