+++ crates/lib/core/codegen/src/templates/field_constants.masm.tpl
