+++ crates/lib/core/codegen/src/templates/field_extra_ops.masm.tpl
