+++ crates/lib/core/codegen/src/templates/u256_constants.masm.tpl
