+++ crates/lib/core/precompiles/asm/crypto/hashes/keccak256.masm
