+++ crates/lib/core/precompiles/asm/crypto/mod.masm
