+++ crates/lib/core/precompiles/asm/math/curve/mod.masm
