+++ crates/lib/core/src/handlers/precompiles/mod.rs
@@ -1,0 +1,4 @@
+//! Host event handlers for bundled precompile-backed MASM wrappers.
+
+pub mod keccak256;
+pub mod uint_field_inv;
