+++ crates/lib/core/tests/precompiles/u256.rs
@@ -1,12 +1,12 @@
 use miden_precompiles::U256;
 
-use crate::uint_fixtures::assert_u256_contract;
+use super::uint_fixtures::assert_u256_contract;
 
 const MODULE: &str = "u256";
 
 #[test]
 fn u256_satisfies_uint_contract() {
-    let u256_masm = miden_precompiles_codegen::masm::render_u256_masm()
+    let u256_masm = miden_core_lib_codegen::masm::render_u256_masm()
         .expect("u256 MASM generation must succeed");
     assert_u256_contract::<U256>(MODULE, &u256_masm);
 }
