--- precompiles/asm/miden-project.toml
@@ -1,13 +1,0 @@
-[package]
-name = "miden-precompiles"
-version = "0.25.0"
-
-[lib]
-namespace = "miden::precompiles"
-path = "mod.masm"
-
-[profile.release]
-# Always produce debug information, as it can be stripped later by the VM
-debug = true
-# Use workspace-relative file paths in debug info for portability
-trim-paths = true
