+++ precompiles/src/lib.rs
@@ -5,14 +5,7 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
-use alloc::sync::Arc;
-#[cfg(feature = "std")]
-use std::path::PathBuf;
-
-use miden_core::{deferred::PrecompileRegistry, serde::Deserializable};
-use miden_mast_package::Package;
-use miden_processor::HostLibrary;
-use miden_utils_sync::LazyLock;
+use miden_core::deferred::PrecompileRegistry;
 
 mod codec;
 mod hash;
@@ -23,94 +16,25 @@ pub use hash::{HashAssertNode, HashFunction, HashPrecompile, keccak256::Keccak25
 pub use math::{
     curve::{
         CurveCoefficient, CurveId, CurveNodeRef, CurvePoint, CurvePrecompile, CurveSpec, K1_A_PTR,
-        K1_B_PTR, K1_GROUP_PTR, ShortWeierstrassSpec, curve_coefficients,
+        K1_B_PTR, K1_GROUP_PTR, SECP256K1_GENERATOR_X, SECP256K1_GENERATOR_Y, SECP256K1_ID,
+        ShortWeierstrassSpec, curve_coefficients,
     },
     k1_base::K1Base,
     k1_scalar::K1Scalar,
     u256::U256,
     uint::{
-        K1_BASE_BOUND_PTR, K1_SCALAR_BOUND_PTR, Limbs, U256_BOUND_PTR, UintDomain, UintNodeRef,
-        UintPrecompile, UintSpec,
+        K1_BASE_BOUND_PTR, K1_SCALAR_BOUND_PTR, Limbs, ONE_LIMBS, TWO_LIMBS, U256_BOUND_PTR,
+        UintDomain, UintNodeRef, UintPrecompile, UintSpec, ZERO_LIMBS,
     },
 };
-
-#[cfg(feature = "std")]
-#[doc(hidden)]
-pub fn asm_source_dir() -> PathBuf {
-    PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("asm")
-}
-
-// PRECOMPILES LIBRARY
-// ================================================================================================
-
-/// The Miden precompiles library, wrapping the compiled `miden-precompiles` [`Package`].
-///
-/// The package bundles MASM support procedures for hash wrappers, arithmetic wrappers, signature
-/// wrappers, and deferred-DAG helpers. These MASM modules are currently internal implementation
-/// detail for core-library facades and precompile tests, while the crate's
-/// [`PrecompileRegistry`] for deferred evaluation is provided separately via [`registry`].
-///
-/// [`Package`]: miden_mast_package::Package
-#[derive(Clone)]
-pub struct PrecompilesLibrary(Arc<Package>);
-
-impl PrecompilesLibrary {
-    /// Serialized representation of the `miden-precompiles` package.
-    pub const SERIALIZED: &'static [u8] =
-        include_bytes!(concat!(env!("OUT_DIR"), "/assets/miden-precompiles.masp"));
-
-    /// Returns a reference to the underlying [`Arc<Package>`].
-    pub fn package(&self) -> Arc<Package> {
-        self.0.clone()
-    }
-}
-
-impl From<&PrecompilesLibrary> for HostLibrary {
-    fn from(precompiles_lib: &PrecompilesLibrary) -> Self {
-        let mut library = HostLibrary::from(precompiles_lib.package());
-        library.handlers = event_handlers::default_event_handlers();
-        library
-    }
-}
-
-impl Default for PrecompilesLibrary {
-    fn default() -> Self {
-        static PRECOMPILES: LazyLock<PrecompilesLibrary> = LazyLock::new(|| {
-            let contents = Package::read_from_bytes(PrecompilesLibrary::SERIALIZED)
-                .expect("failed to read miden-precompiles package!");
-            PrecompilesLibrary(Arc::new(contents))
-        });
-        PRECOMPILES.clone()
-    }
-}
-
-// EVENT HANDLERS
-// ================================================================================================
-
-pub mod event_handlers {
-    use alloc::{sync::Arc, vec, vec::Vec};
-
-    use miden_core::events::EventName;
-    use miden_processor::event::EventHandler;
-
-    use crate::{hash::handlers as hash_handlers, math::uint::handlers as uint_handlers};
-
-    /// Event used by generated field uint wrappers to request an inverse witness from the host.
-    pub const UINT_FIELD_INV_EVENT_NAME: EventName = uint_handlers::UINT_FIELD_INV_EVENT_NAME;
-
-    /// Returns the default host event handlers required by this precompiles package.
-    pub fn default_event_handlers() -> Vec<(EventName, Arc<dyn EventHandler>)> {
-        vec![
-            hash_handlers::keccak256_digest_event_handler(),
-            uint_handlers::field_inv_event_handler(),
-        ]
-    }
-}
 
 // REGISTRY
 // ================================================================================================
 
 /// Returns a [`PrecompileRegistry`] containing the precompiles provided by this crate.
+///
+/// TODO: If constructing the official registry becomes measurable overhead, consider a
+/// cached/shared registry for default processor initialization.
 pub fn registry() -> PrecompileRegistry {
     PrecompileRegistry::new()
         .with_precompile(Keccak256Precompile::default())
